requirements.txt
Flask
gunicorn
imagesize
itsdangerous
matplotlib
numpy
pytz
pillow
requests